// packages/api/src/schemas/public/SensorData.ts
export default interface SensorData {
  time: Date;

  sensor_id: string;

  battery: number | null;

  temperature: number | null;

  dist_z0: number | null;

  dist_z1: number | null;

  dist_z2: number | null;

  dist_z3: number | null;

  dist_z4: number | null;

  dist_z5: number | null;

  dist_z6: number | null;

  dist_z7: number | null;

  dist_z8: number | null;

  dist_z9: number | null;

  dist_z10: number | null;

  dist_z11: number | null;

  dist_z12: number | null;

  dist_z13: number | null;

  dist_z14: number | null;

  dist_z15: number | null;

  status_z0: number | null;

  status_z1: number | null;

  status_z2: number | null;

  status_z3: number | null;

  status_z4: number | null;

  status_z5: number | null;

  status_z6: number | null;

  status_z7: number | null;

  status_z8: number | null;

  status_z9: number | null;

  status_z10: number | null;

  status_z11: number | null;

  status_z12: number | null;

  status_z13: number | null;

  status_z14: number | null;

  status_z15: number | null;
}